requirements.txt
Flask
Flask-Bcrypt
Flask-Migrate
Flask-RESTX
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-JWT-Extended
Werkzeug
